+++ tests/Mocks/ZohoPropertyListing.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace Asciisd\ZohoV8\Tests\Mocks;
+
+use Asciisd\ZohoV8\Models\ZohoModel;
+
+class ZohoPropertyListing extends ZohoModel
+{
+    protected const MODULE_API_NAME = 'Property_Listings';
+}